Add option to only scroll in one direction when using trackpad
This will pin the scrolling direction to the one at the beginning of a
trackpad scrolling gesture, which helps prevent horizontal drift when
the user is simply trying to scroll vertically up and down. At the
beginning of the next scroll gesture the scrolling direction will be
reset. This is the same as Visual Studio Code's "Scroll Predominant
Axis" setting.