Add release.yml to aid pre-release release notes generation
Currently this is only for pre-release builds as main releases usually
use hand-written release notes. However, eventually main release notes
may use this as part of the full release notes to allow people to look
at the full list of changes, but I haven't decided if that's desired
yet as it may just be duplicate information.