Update xcodeproj version, generate a compatible one for Xcode 8+
Certain newer Xcode features such as parallel builds require a newer
version to be specified in the project file. Currently, even though we
have enabled `BuildIndependentTargetsInParallel` in the project file,
because the objectVersion is still set at 47 (Xcode 6.3), the setting is
not used and we get a warning during build time. Upgrade the version to
54 (as automatically done by Xcode) which is Xcode 12 compatible in
order to take advantage of parallle builds.

One caveat of incrementing the version is that older Xcode versions will
not be able to open the project anymore (despite them being able to
build it in command-line using xcodebuild). Generate a new xcodeproj
with the old objectVersion and have CI check that it's not out of date
whenever we change the original xcodeproj, so people using older Xcode
versions can still debug and use Xcode with MacVim.

Note that the compatible version is compatible with Xcode 6.4, but the
xib files are only buildable in Xcode 8+, so we just call it
MacVim_xcode8.xcodeproj.

Related: #1100