Fix macos-12 CI runner not being able to build gettext
When building gettext, Homebrew tries to upgrade all dependencies, and
Python is somehow a downstream dependency and it fails link in certain
binaries like 2to3 which GitHub's CI environment already installed
separately. This is a messy situation in general, but to fix it, just
set `HOMEBREW_NO_INSTALLED_DEPENDENTS_CHECK` so Homebrew doesn't go
helpfully upgrade our dependencies when all we want is to build the
package itself. For the most part it should not cause any issues.