ci: Fix Python 2 in CI stalling tests causing flaky failures
In MacVim CI, we install Python 2 for backwards compatibility reasons,
but CI tests prefer Python 2 so we end up launching that instead, and it
seems like that was the cause for some flaky failures (e.g.
`Test_terminal_aucmd_on_close`). This is probably due to the fact that
Python 2 is only availble in x86 and launching it on an ARM Mac needs to
go through Rosetta which takes time and fails the timeout.