Add setting for whether to show document icon at title bar
Don't automatically hide the icon while in transparent title bar mode,
since it's a little confusing to the user. Instead, just use an extra
setting for that.

Also, when hiding the icon, properly hide it now by clearing it.
Previously it was just making an invisible icon which was still
draggable and didn't behave properly. Make sure after we set the setting
it's refreshed correctly.

Fix #1503