Fix non-native fullscreen quit to black screen bug
A previous change #1510 introduced a piece of debug code that
erroneously overrode the `window` property to always return the normal
window whereas in normal behavior it should return the full screen
window when in non-native full screen. This property was only added for
supporting unit tests and also completely unnecessary because the
superclass `NSWindowController` already provides it (the one that we
accidentally overrode). Just remove this.

Fix #1515