Fix resizing MacVim window occasionally result in a stale wrong Vim size
First issue was that if you resize MacVim quickly (could be simulated by
stalling Vim by calling `:sleep 5`), the existing logic for caching the
size was wrong. It was using `maxRows`/`maxColumns` from the text view
to decide if we send an updated message to Vim, but those are only
updated after Vim has responded back to us. That means if we have two
resize events before Vim could respond, that could lead to wrong results.
One example would be quickly resizing Vim, then resizing it back to
original size. To fix this, add a new "pending" rows/cols whenever we
update Vim, and use that for caching instead as it always reflects what
we want Vim's size to be.

Another issue was that in live resizing, if the user resizes Vim quickly
(which again could be simulated by stalling Vim), the rate limiting
logic would not send Vim with the latest size until the live event has
finished (the user let go of the mouse button). This feels weird when it
happens. Instead, fix it so that whenever we have rate limited the IPC
commands, we immediately send the updated resize message once Vim has
handled the last one, instead of waiting to do it in `liveResizeDidEnd`.