Fix non-native full screen bad interaction with window delegate
This issue was detected when adding new tests for non-native full
screen. When setting 'fuoptions' to empty and going full screen, the
code would fail but also crash, due to a number of issues:

1. `enterFullScreen`'s setting of presentationOptions somehow triggers a
   window resize on the normal window, which leads to `windowDidResize`
   delegate being called. This is a degenerate situation as the window
   controller thinks we are already in full screen even though we
   haven't finished setting up yet, and in particular `nonFuVimViewSize`
   is still 0. This leads to the desired desired frame size being set
   incorrectly to 0.
2. `constrainRows:columns:toSize` does not sanity check the results when
   we have such degenerate small sizes, and end up calculating a desired
   rows to -1, which makes no sense. This leads to various wrong
   calculations.
3. MMCoreTextView loops through the grid using a size_t even though
   grid.rows is an int. This is a poor code practice in general and
   results in a crash as the loop comparison cast the -1 to size_t and
   the loop didn't correctly terminate.

Fix 1 (the root cause) by making sure we detact the window delegate
immediately when entering full screen to prevent any stray window
messages from causing issues. Also safeguard `windowDidResize` so it
only handles the full screen path if we have `fullScreenEnabled` set.

For 2 and 3, add the sanity checks and also fix the size_t to use int
when looping.

For some reason this issue only showed up in CI but not in local
testing, probably due to different screen environments.