New defaults: Smoothly resize window, guioptions+=k
Change MacVim defaults to smoothly resize the window (instead of setting
a resize constraint based on the grid size). Main motivation for this
change is that this aligns to how most modern terminal emulators and
Neovim GUIs and text editors work, and MacVim is implemented to still
look nice while being smoothly resized. It also works better with the
window manager. In particular, macOS 15 Sequoia introduced window tiling
that allows you to pin a window to a side/corner of the screen. When a
window is tiled, macOS actually actively ignores the resize constraint
and allows the user to resize the window smoothly even if the window
otherwise cannot be. There were also previous bugs with exiting native
full screen and third-party snapping tools not working well with
non-smooth resize. It's simply better to just default to this more
intuitive option.
- To get old behavior back: Go to "Settings…" and uncheck "Smoothly
  resizes window" under "General".

Add "k" to `guioptions` setting in Vim. This works well with smooth
resize and has similar rationale. It makes it so that when
adding/removing a scrollbar or tab bar, MacVim wouldn't resize itself.
Instead, it resizes the Vim window within it to accommodate. This again
aligns more with how most apps work, and works better when tiled to a
side of the screen.
- To get old behavior back: Add `set guioptions-=k` to vimrc.

See #1277 for more discussions on defaults.