ci: Don't use guioptions-k in Vim tests, clean up defaults for gui tests
This keeps it consistent with other Vim GUIs and prevent unintentional
breakages, since we have enabled guioptions-k as default now in MacVim
which isn't the case for other platforms.

Also, since MacVim tests are run before Vim GUI tests, there's a
possibility that there are stale auto-saved states leftover that
interferes with them. Usually they should not matter, but we clean them
up anyway to make sure we have a consistent state to run tests from.