MMTabline: Clear all tabs while showing / hiding tab line
When the GUI tab line is hidden (either due to `set showtabline` or `set
go-=e`), Vim stops sending update messages to the GUI. This means the
tabs were stale and would never get cleared, and the moment we show the
tabs we see a confusing animation when the tabs quicly try to rearrange
to match Vim's state. To fix this, just make sure to clean up and remove
all of them when we hide the tab line.